% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_crypto.R
\name{get_market_crypto}
\alias{get_market_crypto}
\title{Get market crypto data from FMP cloud}
\usage{
get_market_crypto(
  symbol,
  multiply = 1,
  time = "hour",
  from = as.character(Sys.Date() - 7),
  to = as.character(Sys.Date()),
  api_key,
  limit
)
}
\arguments{
\item{symbol}{Crypto symbol}

\item{multiply}{Multiply time yo get final frequency}

\item{time}{Time frequency}

\item{from}{Start date, character}

\item{to}{End date, character}

\item{api_key}{FMP cloud API key}

\item{limit}{= Response limit, default is 500. (Maximum 1000)}
}
\value{
Data table of OHLCV prices
}
\description{
Get market crypto data from FMP cloud API.
}
\examples{
get_market_crypto(symbol = 'btcusd', multiply = 1, time = 'h', from = as.character(Sys.Date() - 7),
                  to = as.character(Sys.Date()), api_key = Sys.getenv("APIKEY"), limit = 100)
}
