% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markt_equities.R
\name{get_market_equities}
\alias{get_market_equities}
\title{Get market data from FMP cloud}
\usage{
get_market_equities(
  symbol,
  multiply = 1,
  time = "hour",
  from = as.character(Sys.Date() - 7),
  to = as.character(Sys.Date()),
  api_key
)
}
\arguments{
\item{symbol}{Stock symbol}

\item{multiply}{Multiply time yo get final frequency}

\item{time}{Time frequency}

\item{from}{Start date, character}

\item{to}{End date, character}

\item{api_key}{FMP cloud API key}
}
\value{
Data table of OHLCV prices
}
\description{
Get market data from FMP cloud API.
}
\examples{
get_market_equities("AAPL", api_key = Sys.getenv("APIKEY"))
}
