% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radf_point.R
\name{radf_point}
\alias{radf_point}
\title{Point radf}
\usage{
radf_point(symbols, end_date, window, price_lag, use_log, api_key, ...)
}
\arguments{
\item{symbols}{Stock symbols}

\item{end_date}{End date.}

\item{window}{Window size}

\item{price_lag}{Number of price lags to use}

\item{use_log}{Should log of prices be used}

\item{api_key}{FMP cloud API key}

\item{...}{arguments for get_market_equities function}
}
\value{
Radf values; output from exuber package
}
\description{
Calcualte radf values for point in time.
}
