% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_radf.R
\name{roll_radf}
\alias{roll_radf}
\title{Exuber on rolling window}
\usage{
roll_radf(price, use_log, window, price_lags, no_cores = 4L)
}
\arguments{
\item{price}{Stock price}

\item{use_log}{Should log of prices be used}

\item{window}{Window size}

\item{price_lags}{Number of price lags to use}

\item{no_cores}{Number of cores to use}
}
\value{
Radf values; output from exuber package
}
\description{
Calculate radf values on rolling window using exuber apckage
}
